complete the agent to move to the center of the object
에이전트가 물체의 중심으로 이동하도록 구현하였다. 모든 복셀에 대한 엣지 코스트를 0~10으로 매핑하여 튀는 현상을 없앨 수 있었다.

+++ Client/X-Engine/include/Scene.h
@@ -111,7 +111,7 @@ public:
 	VoxelState		GetVoxelState(const Pos& index) const;
 	VoxelCondition	GetVoxelCondition(const Pos& index) const;
 	int				GetProximityCost(const Pos& index) const;
-	int				GetEdgeCost(const Pos& index, bool isRowEdge) const;
+	float			GetEdgeCost(const Pos& index, bool isRowEdge) const;
 	void			SetVoxelState(const Pos& index, VoxelState state) const;
 	void			SetVoxelCondition(const Pos& index, VoxelCondition condition) const;
 	void			SetProximityCost(const Pos& index, int cost) const;
